+++ sdk/storage/Azure.Storage.Blobs/src/BlobServiceClient.cs
@@ -72,7 +72,7 @@ namespace Azure.Storage.Blobs
         /// <summary>
         /// Gets the Storage account name corresponding to the service client.
         /// </summary>
-        public string AccountName
+        public virtual string AccountName
         {
             get
             {